+++ src/lua/EntityApi.cpp
@@ -3529,6 +3529,10 @@ int LuaContext::hero_api_freeze(lua_State* l) {
   return state_boundary_handle(l, [&] {
     Hero& hero = *check_hero(l, 1);
 
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
+
     hero.start_frozen();
 
     return 0;
@@ -3544,6 +3548,10 @@ int LuaContext::hero_api_unfreeze(lua_State* l) {
 
   return state_boundary_handle(l, [&] {
     Hero& hero = *check_hero(l, 1);
+
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
 
     hero.start_state_from_ground();
 
@@ -3564,6 +3572,10 @@ int LuaContext::hero_api_walk(lua_State* l) {
     bool loop = LuaTools::opt_boolean(l, 3, false);
     bool ignore_obstacles = LuaTools::opt_boolean(l, 4, false);
 
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
+
     hero.start_forced_walking(path, loop, ignore_obstacles);
 
     return 0;
@@ -3581,6 +3593,9 @@ int LuaContext::hero_api_start_attack(lua_State* l) {
     Hero& hero = *check_hero(l, 1);
 
     if (hero.can_start_sword()) {
+      if (hero.get_state()->is_stopping()) {
+        LuaTools::error(l, "Cannot change state: another state is already about to start");
+      }
       hero.start_sword();
     }
 
@@ -3601,6 +3616,9 @@ int LuaContext::hero_api_start_attack_loading(lua_State* l) {
 
     if (hero.can_start_sword()) {
       hero.start_sword_loading(spin_attack_delay);
+      if (hero.get_state()->is_stopping()) {
+        LuaTools::error(l, "Cannot change state: another state is already about to start");
+      }
     }
 
     return 0;
@@ -3623,6 +3641,10 @@ int LuaContext::hero_api_start_item(lua_State* l) {
           std::string("Cannot use item '" + item.get_name() + "': this item is not saved"));
     }
     if (hero.can_start_item(item)) {
+      if (hero.get_state()->is_stopping()) {
+        LuaTools::error(l, "Cannot change state: another state is already about to start");
+      }
+
       hero.start_item(item);
     }
 
@@ -3641,6 +3663,9 @@ int LuaContext::hero_api_start_grabbing(lua_State* l) {
     Hero& hero = *check_hero(l, 1);
 
     if (hero.get_equipment().has_ability(Ability::GRAB)) {
+      if (hero.get_state()->is_stopping()) {
+        LuaTools::error(l, "Cannot change state: another state is already about to start");
+      }
       hero.start_grabbing();
     }
 
@@ -3660,6 +3685,10 @@ int LuaContext::hero_api_start_jumping(lua_State* l) {
     int direction = LuaTools::check_int(l, 2);
     int length = LuaTools::check_int(l, 3);
     bool ignore_obstacles = LuaTools::opt_boolean(l, 4, false);
+
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
 
     hero.start_jumping(direction, length, ignore_obstacles, false);
 
@@ -3701,6 +3730,9 @@ int LuaContext::hero_api_start_treasure(lua_State* l) {
 
     const ScopedLuaRef& callback_ref = LuaTools::opt_function(l, 5);
 
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
     hero.start_treasure(treasure, callback_ref);
 
     return 0;
@@ -3717,6 +3749,10 @@ int LuaContext::hero_api_start_victory(lua_State* l) {
   return state_boundary_handle(l, [&] {
     Hero& hero = *check_hero(l, 1);
     ScopedLuaRef callback_ref = LuaTools::opt_function(l, 2);
+
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
 
     hero.start_victory(callback_ref);
 
@@ -3738,6 +3774,10 @@ int LuaContext::hero_api_start_boomerang(lua_State* l) {
     const std::string& tunic_preparing_animation = LuaTools::check_string(l, 4);
     const std::string& sprite_name = LuaTools::check_string(l, 5);
 
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
+
     hero.start_boomerang(max_distance, speed,
         tunic_preparing_animation, sprite_name);
 
@@ -3755,6 +3795,10 @@ int LuaContext::hero_api_start_bow(lua_State* l) {
   return state_boundary_handle(l, [&] {
     Hero& hero = *check_hero(l, 1);
 
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
+
     hero.start_bow();
 
     return 0;
@@ -3771,6 +3815,10 @@ int LuaContext::hero_api_start_hookshot(lua_State* l) {
   return state_boundary_handle(l, [&] {
     Hero& hero = *check_hero(l, 1);
 
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
+
     hero.start_hookshot();
 
     return 0;
@@ -3786,6 +3834,10 @@ int LuaContext::hero_api_start_running(lua_State* l) {
 
   return state_boundary_handle(l, [&] {
     Hero& hero = *check_hero(l, 1);
+
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot change state: another state is already about to start");
+    }
 
     hero.start_running();
 
@@ -3850,6 +3902,10 @@ int LuaContext::hero_api_start_state(lua_State* l) {
     if (state->is_current_state()) {
       LuaTools::arg_error(l, 1, "This state is already active");
     }
+    if (hero.get_state()->is_stopping()) {
+      LuaTools::error(l, "Cannot start custom state: another state is already about to start");
+    }
+
     hero.start_custom_state(state);
 
     return 0;