+++ tests/cmake/AddTestMaps.cmake
@@ -144,6 +144,7 @@ list(APPEND LUA_TEST_MAPS
   "bugs/1268_movement_ignore_suspend_in_timer"
   "bugs/1275_map_set_tileset"
   "bugs/1286_crash_state_set_can_traverse"
+  "bugs/1290_crash_state_on_finished"
   "bugs/1291_hero_start_attack_loading"
   "bugs/1297_target_movement_speed"
   "bugs/1301_state_set_can_be_hurt_function"