+++ changelog.md
@@ -10,6 +10,7 @@
 * Fix crash when setting events on a joypad after the game was reset (#1898).
 * Fix crash when setting an ability in game:on_started() (#1920).
 * Fix crash when calling `hero:get_map()` from `game:on_started()` (#1937).
+* Fix crash when attempting to change hero state during state:on_finished() (#1290).
 * Fix `game:set_command_joypad_binding()` with old button identifiers (#1919).
 * Fix `music:set_music_channel_volume` crashes the engine if channel number is 0 (#1924).
 * Fix diagonal jumpers preventing interactions of entities overlapping them (#1928).